\name{fmolsR-package}
\alias{fmolsR-package}
\alias{fmolsR}
\title{Training scaffold for reproducing EViews FMOLS in R}
\description{
This package directory is a training scaffold for learning how to build an R
implementation of EViews FMOLS estimators with benchmark-driven validation.
}
\details{
The finished estimator code is intentionally not included yet. Use the sibling
training workspace documentation to implement the package in stages.
}
\docType{package}
\keyword{package}
